# apis/rest/api-server/openapi.yaml
openapi: 3.0.3
info:
  title: API Gateway — API Server HTTP API
  version: 1.3.0
  description: |
    HTTP API for the **API Server** control plane: authentication material (JWT, JWKS), contract export,
    registry views backed by **etcd** (controllers, tenants, bundles), and operational endpoints.

    **Data model:** registry responses are **read-only snapshots** of stored state. Environment configuration
    lifecycle is owned by the **Gateway Controller**; this API does not expose CRUD for environments.

    **JSON naming:** request and response bodies use **snake_case** field names unless noted.

    **v1.1.0** added optional `config_source` on static bundles, `sources_fingerprint` and `effective_generation` on
    environments. **v1.2.0** added `environment_config_source`, `services` (static upstreams with per-line
    `config_source`). **v1.3.0** nests that observability in **`meta`** on each entity (`provenance.config_source`, plus
    `effective_generation` / `sources_fingerprint` only on environment `meta`); flat `config_source` fields on bundles
    and services are removed in favor of `meta.provenance`.

    **Errors:** all failed requests use **`application/problem+json`** (RFC 9457 / RFC 7807 Problem Details).
  contact:
    name: Merionyx API Gateway
  license:
    name: Proprietary
    url: https://merionyx.com/
servers:
- url: '{baseUrl}'
  description: Configurable base URL (set `baseUrl` when using this spec in clients).
  variables:
    baseUrl:
...
security: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Reserved for future use. The current API Server revision does not validate Bearer tokens on these routes.
  parameters:
    Limit:
      name: limit
      in: query
      description: Maximum number of items to return (cursor pagination).
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    Cursor:
      name: cursor
      in: query
      description: Opaque cursor from a previous response (`next_cursor`). Omit for the first page.
      required: false
      schema:
        type: string
...
  responses:
...
    NotModified:
      description: Not Modified — representation unchanged (use with `If-None-Match`).
...
paths:
...
  /api/v1/controllers:
    get:
      tags: [ Registry ]
      operationId: listControllers
      summary: List registered controllers
      description: |
        Lists Gateway Controllers registered in etcd (`controller_id`, `tenant`, declared environments and static bundles). Paginated with `limit` / `cursor`. Use `If-None-Match` with the previous `ETag` to obtain `304` when nothing changed.
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Paginated controllers
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllerListResponse'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/controllers/{controller_id}:
    get:
      tags: [ Registry ]
      operationId: getController
      summary: Get controller by ID
      description: |
        Returns one controller record by `controller_id`. Supports conditional GET via `If-None-Match` / `ETag`. `404` if the controller is unknown.
      parameters:
      - $ref: '#/components/parameters/ControllerId'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Controller
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Controller'
        '304':
          $ref: '#/components/responses/NotModified'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/controllers/{controller_id}/heartbeat:
    get:
      tags: [ Registry ]
      operationId: getControllerHeartbeat
      summary: Get last heartbeat for a controller
      description: |
        Returns the last heartbeat timestamp stored for the controller (operator visibility). Not the same payload as full controller registration; use when you only need freshness. `404` if the controller does not exist.
      parameters:
      - $ref: '#/components/parameters/ControllerId'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Heartbeat record
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllerHeartbeat'
        '304':
          $ref: '#/components/responses/NotModified'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/tenants:
    get:
      tags: [ Registry ]
      operationId: listTenants
      summary: List distinct tenants
      description: |
        Returns distinct `tenant` values observed across registered controllers. Cursor-paginated; combine with registry list endpoints to drill down.
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Paginated tenant names
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TenantListResponse'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/tenants/{tenant}/controllers:
    get:
      tags: [ Registry ]
      operationId: listControllersByTenant
      summary: List controllers for a tenant
      description: |
        Lists controllers whose `tenant` path parameter matches. Same response shape as `GET /api/v1/controllers` but scoped to one tenant.
      parameters:
      - $ref: '#/components/parameters/Tenant'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Paginated controllers
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllerListResponse'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/tenants/{tenant}/environments:
    get:
      tags: [ Registry ]
      operationId: listEnvironmentsByTenant
      summary: Aggregated environments observed for a tenant (read-only snapshot)
      description: |
        Read-only aggregated view of environments (and their static bundles) as reported for this tenant. Environment lifecycle is owned by the Gateway Controller; this endpoint does not mutate configuration.
      parameters:
      - $ref: '#/components/parameters/Tenant'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Paginated environments
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EnvironmentListResponse'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/tenants/{tenant}/bundles:
    get:
      tags: [ Registry ]
      operationId: listBundlesByTenant
      summary: Static bundles aggregated across environments for a tenant
      description: |
        Flattens static bundle descriptors (`name`, `repository`, `ref`, `path`) across environments for the tenant. Useful for overview UIs; not a substitute for etcd bundle-key APIs.
      parameters:
      - $ref: '#/components/parameters/Tenant'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Paginated bundle descriptors
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundleRefListResponse'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/bundles:
    get:
      tags: [ Bundles ]
      operationId: listBundleKeys
      summary: List bundles present in etcd
      description: |
        Lists bundles that exist under the API Server’s contract snapshot prefix in etcd (`repository`, `ref`, `path`, canonical `key`). Paginated; the `key` may require percent-encoding when used as path segments elsewhere.
      parameters:
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Paginated bundle descriptors (etcd snapshot roots)
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BundleRefListResponse'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/bundles/contracts:
    get:
      tags: [ Bundles ]
      operationId: listContractsInBundle
      summary: List contract names for a bundle key
      description: |
        Returns logical contract names stored for the bundle in etcd. Identify the bundle either with
        query parameter `bundle_key` (full key as in list bundle keys) **or** with `repo` and `ref`
        (and optional `path`) — same composition as elsewhere in this API. Paginated list; use
        `GET .../contracts/{contract_name}` for full document bodies.
      parameters:
      - $ref: '#/components/parameters/BundleKeyQuery'
      - $ref: '#/components/parameters/BundleRepoQuery'
      - $ref: '#/components/parameters/BundleRefQuery'
      - $ref: '#/components/parameters/BundlePathQuery'
      - $ref: '#/components/parameters/Limit'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Paginated contract names
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractNameListResponse'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/bundles/contracts/{contract_name}:
    get:
      tags: [ Bundles ]
      operationId: getContractInBundle
      summary: Get a single contract document for a bundle
      description: |
        Returns one contract document from etcd-backed storage for the bundle (JSON object; YAML sources may be normalized to JSON). Identify the bundle with `bundle_key` **or** `repo` + `ref` (+ optional `path`) query parameters, same as `listContractsInBundle`. Supports `If-None-Match` for caching. `404` if the bundle or contract is missing.
      parameters:
      - $ref: '#/components/parameters/ContractName'
      - $ref: '#/components/parameters/BundleKeyQuery'
      - $ref: '#/components/parameters/BundleRepoQuery'
      - $ref: '#/components/parameters/BundleRefQuery'
      - $ref: '#/components/parameters/BundlePathQuery'
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Contract document (JSON)
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContractDocument'
        '304':
          $ref: '#/components/responses/NotModified'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/v1/status:
    get:
      tags: [ System ]
      operationId: getStatus
      summary: Dependency status for operators
      description: |
        Rich snapshot for dashboards and alerting (API Server self-check, etcd, Contract Syncer). **Not** a Kubernetes probe — use `GET /ready` for orchestrator readiness. Exact fields may evolve. Conditional GET via `ETag` / `If-None-Match`.
      parameters:
      - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Status snapshot
          headers:
            ETag:
              $ref: '#/components/headers/ETag'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '304':
          $ref: '#/components/responses/NotModified'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'